+++ src/app/store/[storeId]/page.tsx
